Use a double to store state of LabelledSlider
The LabelledSlider is used to change the depth setting. Some
touchpads (e.g. on macOS) can send dozens or hundreds of scroll
events per second, each with a very small delta. Unless we keep
track of the "true" state of the slider, these deltas will be lost
when they are rounded into the integer depth setting.

Fixes #287

+++ mapview.h
@@ -69,8 +69,8 @@ class MapView : public QWidget {
 
  signals:
   void hoverTextChanged(QString text);
-  void demandDepthChange(int value);
-  void demandDepthValue(int value);
+  void demandDepthChange(double value);
+  void demandDepthValue(double value);
   void showProperties(QVariant properties);
   void addOverlayItemType(QString type, QColor color);
   void coordinatesChanged(int x, int y, int z);
